+++ crates/store/src/server/block_producer.rs
@@ -1,7 +1,7 @@
 use std::convert::Infallible;
 
 use miden_crypto::dsa::ecdsa_k256_keccak::Signature;
-use miden_node_proto::decode::GrpcDecodeExt;
+use miden_node_proto::decode::{GrpcDecodeExt, read_account_id, read_account_ids};
 use miden_node_proto::domain::proof_request::BlockProofRequest;
 use miden_node_proto::errors::ConversionError;
 use miden_node_proto::generated::store::block_producer_server;
@@ -20,8 +20,6 @@ use crate::errors::ApplyBlockError;
 use crate::server::api::{
     StoreApi,
     conversion_error_to_status,
-    read_account_id,
-    read_account_ids,
     read_block_numbers,
     validate_note_commitments,
     validate_nullifiers,
