+++ crates/store/src/server/ntx_builder.rs
@@ -2,6 +2,7 @@ use std::collections::BTreeSet;
 use std::num::{NonZero, TryFromIntError};
 
 use miden_crypto::merkle::smt::SmtProof;
+use miden_node_proto::decode::{read_account_id, read_block_range, read_root};
 use miden_node_proto::domain::account::AccountInfo;
 use miden_node_proto::errors::ConversionError;
 use miden_node_proto::generated as proto;
@@ -23,14 +24,7 @@ use crate::errors::{
     GetNoteScriptByRootError,
     GetWitnessesError,
 };
-use crate::server::api::{
-    StoreApi,
-    internal_error,
-    invalid_argument,
-    read_account_id,
-    read_block_range,
-    read_root,
-};
+use crate::server::api::{StoreApi, internal_error, invalid_argument};
 
 // NTX BUILDER ENDPOINTS
 // ================================================================================================
